+++ src/main/java/com/cis/batch33/library/model/Member.java
@@ -2,17 +2,17 @@ package com.cis.batch33.library.model;
 
 public class Member {
 
-    private long memberId;
+    private Integer memberId;
     private String firstName;
     private String lastName;
     private String emailAddress;
     private String phoneNumber;
 
-    public long getMemberId() {
+    public Integer getMemberId() {
         return memberId;
     }
 
-    public void setMemberId(long memberId) {
+    public void setMemberId(Integer memberId) {
         this.memberId = memberId;
     }
